feat: [Detail] 일지 삭제 수정사항 반영

+++ lib/domain/usecase/delete_history_usecase.dart
@@ -1,26 +1,41 @@
 // Package imports:
+// Project imports:
+import 'package:bladderly/core/network_checker/network_checker.dart';
+import 'package:bladderly/domain/model/history_status.dart';
+import 'package:bladderly/domain/repository/history_repository.dart';
 import 'package:dartz/dartz.dart';
 import 'package:injectable/injectable.dart';
 
-// Project imports:
-import 'package:bladderly/domain/repository/history_repository.dart';
-
 @lazySingleton
 class DeleteHistoryUsecase {
-  const DeleteHistoryUsecase({required HistoryRepository historyRepository}) : _historyRepository = historyRepository;
+  const DeleteHistoryUsecase({
+    required HistoryRepository historyRepository,
+    required NetworkChecker networkChecker,
+  })  : _historyRepository = historyRepository,
+        _networkChecker = networkChecker;
 
   final HistoryRepository _historyRepository;
+  final NetworkChecker _networkChecker;
 
-  Either<Exception, void> call({
+  Future<Either<Exception, void>> call({
+    required String userId,
     required int historyId,
-  }) {
+  }) async {
     try {
-      final history = _historyRepository.deleteHistoryById(historyId);
-      return Right(history);
-    } on Exception catch (e) {
-      return Left(e);
+      final history = _historyRepository.getHistoryById(historyId);
+
+      if (history == null) return const Right(null);
+
+      await _historyRepository.saveHistory(history.setStatus(HistoryStatus.pending).setDeletedAt(DateTime.now()));
+
+      if (await _networkChecker.isConnected) {
+        await _historyRepository.deleteHistoryByRecordTimeFromServer(userId: userId, recordTime: history.recordTime);
+        _historyRepository.deleteHistoryById(historyId);
+      }
+
+      return const Right(null);
     } catch (e) {
-      return Left(Exception(e));
+      return Left(e is Exception ? e : Exception(e.toString()));
     }
   }
 }
